refactor: move global error page
