+++ app/manage-promoters/[id]/edit/error.tsx
@@ -1,4 +1,4 @@
 "use client"
 
-import ErrorPage from '@/error'
+import ErrorPage from '@/app/error'
 export default ErrorPage
